Separated zod validator files
