+++ src/zod/models/GenericZodSchemaMap.ts
@@ -1,0 +1,6 @@
+import { ZodTypeAny } from "zod";
+import { RequestFields } from "./RequestFields";
+
+export type GenericZodSchemaMap = {
+  [Key in RequestFields]?: ZodTypeAny;
+};
