+++ src/zod/models/RequestFields.ts
@@ -1,0 +1,5 @@
+import { ZodCompatibleRequest } from "./ZodCompatibleRequest";
+
+/* eslint-enable @typescript-eslint/no-explicit-any */
+
+export type RequestFields = keyof ZodCompatibleRequest;
