+++ src/zod/models/ZodCompatibleRequest.ts
@@ -1,0 +1,9 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+
+export interface ZodCompatibleRequest {
+  body?: any;
+  params?: any;
+  query?: any;
+  headers?: any;
+  cookies?: any;
+}
